+++ src/yw2oxmllib/docx/docx_brief_synopsis.py
@@ -19,10 +19,10 @@ class DocxBriefSynopsis(DocxFile):
 <w:p><w:pPr><w:pStyle w:val="Subtitle"/></w:pPr><w:r><w:t>$AuthorName</w:t></w:r></w:p>
 '''
 
-    _partTemplate = '''<w:p><w:pPr><w:pStyle w:val="Heading1"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
+    _partTemplate = '''<w:p><w:pPr><w:pStyle w:val="heading1"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
 '''
 
-    _chapterTemplate = '''<w:p><w:pPr><w:pStyle w:val="Heading2"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
+    _chapterTemplate = '''<w:p><w:pPr><w:pStyle w:val="heading2"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
 '''
 
     _sceneTemplate = '''<w:p><w:pPr><w:pStyle w:val="BodyText"/><w:rPr></w:rPr></w:pPr><w:r><w:rPr></w:rPr><w:t>$Title</w:t></w:r></w:p>