+++ src/yw2oxmllib/docx/docx_export.py
@@ -16,10 +16,10 @@ class DocxExport(DocxFile):
 <w:p><w:pPr><w:pStyle w:val="Subtitle"/></w:pPr><w:r><w:t>$AuthorName</w:t></w:r></w:p>
 '''
 
-    _partTemplate = '''<w:p><w:pPr><w:pStyle w:val="Heading1"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
+    _partTemplate = '''<w:p><w:pPr><w:pStyle w:val="heading1"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
 '''
 
-    _chapterTemplate = '''<w:p><w:pPr><w:pStyle w:val="Heading2"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
+    _chapterTemplate = '''<w:p><w:pPr><w:pStyle w:val="heading2"/></w:pPr><w:r><w:t>$Title</w:t></w:r></w:p>
 '''
 
     _sceneTemplate = '''<w:p><w:pPr><w:pStyle w:val="BodyText"/></w:pPr><w:r><w:t xml:space="preserve">$SceneContent</w:t></w:r></w:p>
@@ -28,7 +28,7 @@ class DocxExport(DocxFile):
     _appendedSceneTemplate = '''<w:p><w:pPr><w:pStyle w:val="BodyTextIndent"/></w:pPr><w:r><w:t xml:space="preserve">$SceneContent</w:t></w:r></w:p>
 '''
 
-    _sceneDivider = '<w:p><w:pPr><w:pStyle w:val="Heading4"/><w:ind w:hanging="0"/><w:rPr></w:rPr></w:pPr><w:r><w:rPr></w:rPr><w:t>* * *</w:t></w:r></w:p>\n'
+    _sceneDivider = '<w:p><w:pPr><w:pStyle w:val="heading4"/><w:ind w:hanging="0"/><w:rPr></w:rPr></w:pPr><w:r><w:rPr></w:rPr><w:t>* * *</w:t></w:r></w:p>\n'
     _fileFooter = DocxFile._DOCUMENT_XML_FOOTER
 
     def _get_chapterMapping(self, chId, chapterNumber):